backend/pyproject.toml
[project]
name = "notebot-backend"
version = "0.1.0"
description = "Notebot Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
    "faster-whisper>=1.1.0",
    "openai>=1.60.0",
    "anthropic>=0.42.0",
    "google-genai>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
